+++ app/ExternalFiles/Extensions/Sync_By_File_Type.php
@@ -85,6 +85,11 @@ class Sync_By_File_Type extends Extension_Base {
 		// add settings.
 		add_action( 'init', array( $this, 'add_settings' ), 20 );
 
+		// bail if this extension for sync is not enabled.
+		if ( ! in_array( $this->get_name(), SynchronizationDialog::get_instance()->get_enabled_extensions(), true ) ) {
+			return;
+		}
+
 		// use our own hooks.
 		add_filter( 'efml_sync_configure_form', array( $this, 'extend_sync_form_in_dialog' ), 10, 2 );
 		add_action( 'efml_sync_save_config', array( $this, 'save_sync_config' ), 10, 2 );
@@ -110,7 +115,7 @@ class Sync_By_File_Type extends Extension_Base {
 		$settings_obj = Settings::get_instance()->get_settings_obj();
 
 		// get the settings page.
-		$settings_page = $settings_obj->get_page( \ExternalFilesInMediaLibrary\Plugin\Settings::get_instance()->get_menu_slug() );
+		$settings_page = $settings_obj->get_page( $settings_obj->get_menu_slug() );
 		if ( ! $settings_page instanceof Page ) {
 			return;
 		}
@@ -148,11 +153,6 @@ class Sync_By_File_Type extends Extension_Base {
 	 * @return string
 	 */
 	public function extend_sync_form_in_dialog( string $form, int $term_id ): string {
-		// only add if it is enabled in settings.
-		if ( ! in_array( $this->get_name(), SynchronizationDialog::get_instance()->get_enabled_extensions(), true ) ) {
-			return $form;
-		}
-
 		// get the actual setting for this term.
 		$allowed_file_types = get_term_meta( $term_id, 'eml_sync_file_types', true );
 
@@ -187,11 +187,6 @@ class Sync_By_File_Type extends Extension_Base {
 	public function save_sync_config( array $fields, int $term_id ): void {
 		// check for nonce.
 		if ( isset( $_GET['nonce'] ) && ! wp_verify_nonce( sanitize_text_field( wp_unslash( $_GET['nonce'] ) ), 'eml-nonce' ) ) {
-			return;
-		}
-
-		// only add if it is enabled in settings.
-		if ( ! in_array( $this->get_name(), SynchronizationDialog::get_instance()->get_enabled_extensions(), true ) ) {
 			return;
 		}