+++ app/Plugin/Init.php
@@ -14,6 +14,7 @@ use ExternalFilesInMediaLibrary\ExternalFiles\Proxy;
 use ExternalFilesInMediaLibrary\Plugin\Admin\Admin;
 use ExternalFilesInMediaLibrary\Services\Services;
 use ExternalFilesInMediaLibrary\ThirdParty\ThirdPartySupport;
+use WP_User_Query;
 
 /**
  * Initialize the plugin, connect all together.