+++ changelog.md
@@ -11,10 +11,14 @@
 
 - Small optimizations for the multisite support
 - Changed initialization of the plugin during activation and deactivation for better compatibility with other plugins
+- Optimized the uninstallation of services
+- User-token for Dropbox as now also deleted on uninstallation
 
 ### Fixed
 
 - Fixed missing hint for not existing synchronisation support for local external sources
+- Fixed wrong spelled plugin slugs in some hooks
+- Fixed check for file type and size on sync even if this extensions are not enabled
 
 ## [5.0.3] - 18.04.2026